requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pyyaml>=6.0
numpy>=1.24.0
scipy>=1.10.0
pycocotools>=2.0.7
tqdm>=4.66.0
matplotlib>=3.8.0
Pillow>=10.0.0
thop>=0.1.1.post2209072238
